sync_verify: always set RELEASE_IMAGE_INITIAL for upgrades
This PR changes the behavior of the release-controller to always set the
`RELEASE_IMAGE_INITAL` env var for upgrade release jobs.  Previously, it
would only do this if the job contained the env var in its job spec,
which means that any generated upgrade job would not be able to properly
be run by release-controller. This also means that upgrade jobs
configured to be run by release-controller would not be able to be
rehearsed, as the required empty `RELEASE_IMAGE_INITIAL` in the
rehearsal would cause the ci-operator incorrectly set the pullspec to an
empty string.